fix(bench_fio): repair test_number_of_settings
FAILED tests/bench_fio_test.py::TestFunctions::test_number_of_settings - AssertionError: 32 != 33

+++ bench_fio/benchlib/display.py
@@ -33,7 +33,7 @@ def calculate_duration(settings, tests):
         number_of_tests = number_of_tests/len(settings["target"])
     time_per_test = settings["runtime"]
     if time_per_test:
-        duration_in_seconds = number_of_tests * time_per_test
+        duration_in_seconds = int(number_of_tests * time_per_test)
         duration = str(datetime.timedelta(seconds=duration_in_seconds))
     else:
         duration = None
